fix: 인증번호가 만료되었을 때 response처리

+++ server/src/services/userService.js
@@ -133,6 +133,9 @@ exports.checkToken = async function(email, token){
         const authInfo = await Auth.findOne({
             email: email
         });
+        if(!authInfo){
+            return 3;
+        }
         if(authInfo.token === token){
             await Auth.deleteOne({
                 email: email
